+++ client/factory_test.go
@@ -9,7 +9,6 @@ import (
 	"io/ioutil"
 
 	"github.com/go-kit/log"
-	"github.com/prometheus/client_golang/prometheus"
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/thanos-io/objstore/tracing/opentelemetry"
@@ -39,32 +38,6 @@ func ExampleBucket() {
 	// false
 }
 
-func ExampleInstrumentedBucket() {
-	// Read the configuration file.
-	confContentYaml, err := ioutil.ReadFile("testconf/filesystem.conf.yml")
-	if err != nil {
-		panic(err)
-	}
-
-	// Create a new bucket.
-	bucket, err := NewBucket(log.NewNopLogger(), confContentYaml, "example")
-	if err != nil {
-		panic(err)
-	}
-
-	// Wrap it with instrumentation.
-	bucket = InstrumentedBucket(bucket, prometheus.NewRegistry())
-
-	// Test it.
-	exists, err := bucket.Exists(context.Background(), "example")
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println(exists)
-	// Output:
-	// false
-}
-
 func ExampleTracingBucketUsingOpenTracing() { //nolint:govet
 	// Read the configuration file.
 	confContentYaml, err := ioutil.ReadFile("testconf/filesystem.conf.yml")
@@ -79,7 +52,7 @@ func ExampleTracingBucketUsingOpenTracing() { //nolint:govet
 	}
 
 	// Wrap it with tracing.
-	bucket = opentracing.TracedBucket(bucket)
+	bucket = opentracing.WrapWithTraces(bucket)
 
 	// Test it.
 	exists, err := bucket.Exists(context.Background(), "example")
@@ -105,7 +78,7 @@ func ExampleTracingBucketUsingOpenTelemetry() { //nolint:govet
 	}
 
 	// Wrap it with tracing.
-	bucket = opentelemetry.TracedBucket(bucket, trace.NewNoopTracerProvider().Tracer("bucket"))
+	bucket = opentelemetry.WrapWithTraces(bucket, trace.NewNoopTracerProvider().Tracer("bucket"))
 
 	// Test it.
 	exists, err := bucket.Exists(context.Background(), "example")